Move `CreateStreamObject` to proper location
