+++ LEGO1/omni/include/mxdssource.h
@@ -2,8 +2,7 @@
 #define MXDSSOURCE_H
 
 #include "mxcore.h"
-
-class MxDSBuffer;
+#include "mxdsbuffer.h"
 
 // VTABLE: LEGO1 0x100dc8c8
 // SIZE 0x14
@@ -27,15 +26,26 @@ public:
 		return !strcmp(p_name, MxDSSource::ClassName()) || MxCore::IsA(p_name);
 	}
 
-	virtual MxLong Open(MxULong) = 0;                    // vtable+0x14
-	virtual MxLong Close() = 0;                          // vtable+0x18
-	virtual MxResult ReadToBuffer(MxDSBuffer* p_buffer); // vtable+0x1c
-	virtual MxResult Read(unsigned char*, MxULong) = 0;  // vtable+0x20
-	virtual MxLong Seek(MxLong, MxS32) = 0;              // vtable+0x24
-	virtual MxULong GetBufferSize() = 0;                 // vtable+0x28
-	virtual MxULong GetStreamBuffersNum() = 0;           // vtable+0x2c
-	virtual MxLong GetLengthInDWords();                  // vtable+0x30
-	virtual MxU32* GetBuffer();                          // vtable+0x34
+	virtual MxLong Open(MxULong) = 0; // vtable+0x14
+	virtual MxLong Close() = 0;       // vtable+0x18
+
+	// FUNCTION: LEGO1 0x100bffd0
+	virtual MxResult ReadToBuffer(MxDSBuffer* p_buffer)
+	{
+		return Read(p_buffer->GetBuffer(), p_buffer->GetWriteOffset());
+	} // vtable+0x1c
+
+	virtual MxResult Read(unsigned char*, MxULong) = 0; // vtable+0x20
+	virtual MxLong Seek(MxLong, MxS32) = 0;             // vtable+0x24
+	virtual MxULong GetBufferSize() = 0;                // vtable+0x28
+	virtual MxULong GetStreamBuffersNum() = 0;          // vtable+0x2c
+
+	// FUNCTION: LEGO1 0x100bfff0
+	virtual MxLong GetLengthInDWords() { return m_lengthInDWords; } // vtable+0x30
+
+	// FUNCTION: LEGO1 0x100c0000
+	virtual MxU32* GetBuffer() { return m_pBuffer; } // vtable+0x34
+
 	MxLong GetPosition() const { return m_position; }
 
 protected:
