+++ LEGO1/omni/include/mxutilities.h
@@ -76,7 +76,6 @@ void OmniError(const char* p_message, MxS32 p_status);
 void SetOmniUserMessage(void (*p_omniUserMessage)(const char*, MxS32));
 MxBool ContainsPresenter(MxCompositePresenterList& p_presenterList, MxPresenter* p_presenter);
 void FUN_100b7220(MxDSAction* p_action, MxU32 p_newFlags, MxBool p_setFlags);
-MxDSObject* CreateStreamObject(MxDSFile*, MxS16);
 MxBool KeyValueStringParse(char*, const char*, const char*);
 
 #endif // MXUTILITIES_H
