+++ LEGO1/omni/src/action/mxdsobject.cpp
@@ -3,6 +3,7 @@
 #include "mxdsaction.h"
 #include "mxdsanim.h"
 #include "mxdsevent.h"
+#include "mxdsfile.h"
 #include "mxdsmediaaction.h"
 #include "mxdsmultiaction.h"
 #include "mxdsobjectaction.h"
@@ -226,4 +227,40 @@ MxDSObject* DeserializeDSObjectDispatch(MxU8*& p_source, MxS16 p_flags)
 	}
 
 	return obj;
+}
+
+// FUNCTION: LEGO1 0x100c0280
+MxDSObject* CreateStreamObject(MxDSFile* p_file, MxS16 p_ofs)
+{
+	MxU8* buf;
+	_MMCKINFO tmpChunk;
+
+	if (p_file->Seek(((MxLong*) p_file->GetBuffer())[p_ofs], SEEK_SET)) {
+		return NULL;
+	}
+
+	if (p_file->Read((MxU8*) &tmpChunk.ckid, 8) == 0 && tmpChunk.ckid == FOURCC('M', 'x', 'S', 't')) {
+		if (p_file->Read((MxU8*) &tmpChunk.ckid, 8) == 0 && tmpChunk.ckid == FOURCC('M', 'x', 'O', 'b')) {
+
+			buf = new MxU8[tmpChunk.cksize];
+			if (!buf) {
+				return NULL;
+			}
+
+			if (p_file->Read(buf, tmpChunk.cksize) != 0) {
+				return NULL;
+			}
+
+			// Save a copy so we can clean up properly, because
+			// this function will alter the pointer value.
+			MxU8* copy = buf;
+			MxDSObject* obj = DeserializeDSObjectDispatch(buf, -1);
+			delete[] copy;
+			return obj;
+		}
+
+		return NULL;
+	}
+
+	return NULL;
 }
