+++ LEGO1/omni/src/stream/mxdsfile.cpp
@@ -8,6 +8,7 @@
 #define SI_MAJOR_VERSION 2
 #define SI_MINOR_VERSION 2
 
+DECOMP_SIZE_ASSERT(MxDSSource, 0x14)
 DECOMP_SIZE_ASSERT(MxDSFile::ChunkHeader, 0x0c)
 DECOMP_SIZE_ASSERT(MxDSFile, 0x7c)
